refactor/#6 [fix] : 알람 기능 테스트 및 에러 수정
* 안드로이드 애뮬레이터 API 31 이상 Error
* 알람 디렉터리 변경
