+++ app/src/main/AndroidManifest.xml
@@ -68,8 +68,9 @@
         </activity>
 
         <service
-            android:name=".worker.FCMService"
-            android:exported="false">
+            android:name=".worker.alarm.FCMService"
+            android:exported="false"
+            android:directBootAware="true">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
